Refatoramento do código do desafio original
