+++ backend/invoice.rb
@@ -1,0 +1,9 @@
+class Invoice
+  attr_reader :billing_address, :shipping_address, :order
+
+  def initialize(attributes = {})
+    @billing_address = attributes.values_at(:billing_address)
+    @shipping_address = attributes.values_at(:shipping_address)
+    @order = attributes.values_at(:order)
+  end
+end
